pyproject.toml
[project]
name = "fastapi-tp6"
version = "1.0.0"
description = "FastAPI Enterprise Edition"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tortoise-orm>=0.20.0",
    "aiosqlite>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.0.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
